+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ## [Unreleased]
 
-## 0.1.0
+## [0.1.0]
 ### Added
 - Added project files.
+- [#7](https://github.com/picimako/just-kitting/issues/7): Added an intention on Java class methods to generate `CallMatcher`
+    initializer calls from them.
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)